Conflicts:
	core/baseLanguage/baseLanguage/languageModels/typesystem.mps
	plugins/vcs/jetbrains/mps/vcs/diff/changes.mps
	core/kernel/smodelRuntime/jetbrains/mps/lang/smodel/generator/smodelAdapter.mps
	...some caches

+++ core/baseLanguage/tuples/solutions/jetbrains.mps.baseLanguage.tuples.sandbox/source_gen/jetbrains/mps/baseLanguage/tuples/sandbox/foo/Fubar.java
@@ -14,6 +14,6 @@ public class Fubar {
   }
 
   public static void main(String[] args) {
-    System.out.println(new Fubar().a.get(1)._0());
+    System.out.println((int) new Fubar().a.get(1)._0());
   }
 }
