+++ core/baseLanguage/tuples/solutions/jetbrains.mps.baseLanguage.tuples.test/source_gen/jetbrains/mps/baseLanguage/tuples/test/Record.java
@@ -44,6 +44,6 @@ public class Record extends MultiTuple._3<String, Integer, Float> {
   }
 
   public String toString() {
-    return this.name() + ":" + this.age() + ":" + this.size();
+    return this.name() + ":" + (int) this.age() + ":" + (float) this.size();
   }
 }
